Move project() and project_path() fixture to global conftest
These are globally useful for all kind of tests.
